v0.10.0
It turns out that poetry has some issues with dependency
management of itself. Therefore, this removes that dependency
and instead assumes that a proper poetry environment is available
if poetry dependencies are being used in the same fashion that pip
and conda are assumed to be available if those types of dependencies
are being used.